Move common structure to its own file
